requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
sentence-transformers
faiss-cpu
numpy
tqdm